requirements-gpu.txt
absl-py==0.7.1
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
imageio==2.5.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
mock==3.0.4
more-itertools==7.0.0
networkx==2.3
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.25.0
Pillow==6.0.0
pluggy==0.9.0
protobuf==3.7.1
py==1.8.0
pyparsing==2.4.0
pytest==4.4.1
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.2.1
selectivesearch==0.4
six==1.12.0
tensorboard==1.13.1
tensorflow-gpu==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
